Queue TXHashes to redis for processing
