+++ internal/transactions/transactions.go
@@ -48,12 +48,13 @@ type Result struct {
 }
 
 func Stream(ctx context.Context, cfg *config.Config, srvc *service.Service) {
+	// Start the streams
 	for _, endpoint := range cfg.Endpoints {
-		go streamTxs(ctx, endpoint, srvc.Logger, *srvc.Redis)
+		go streamEndpoint(ctx, endpoint, srvc.Logger, *srvc.Redis)
 	}
 }
 
-func streamTxs(ctx context.Context, endpoint config.Endpoint, l pkg.Logger, r redis.Client) {
+func streamEndpoint(ctx context.Context, endpoint config.Endpoint, l pkg.Logger, r redis.Client) {
 	conn, resp, err := websocket.Dial(ctx, endpoint.Websocket, nil)
 
 	if err != nil {
@@ -109,7 +110,7 @@ func streamTxs(ctx context.Context, endpoint config.Endpoint, l pkg.Logger, r re
 		l.Info(fmt.Sprintf("Subscription: %s, TxHash: %s", event.Params.Subscrition, event.Params.TxHash))
 
 		// Queue the TX
-		r.Set(context.Background(), "queuedTxHash", event.Params.TxHash, 0)
+		r.RPush(context.Background(), "queuedTxHash", fmt.Sprintf("%s:%s", endpoint.Name, event.Params.TxHash))
 	}
 }
 
@@ -239,5 +240,137 @@ func processTransactionBatch(ctx context.Context, storage *Storage, listName str
 				fmt.Printf("Error storing TX (address: %s, nonce: %s): %v\n", address, nonce, err)
 			}
 		}
+	}
+}
+
+// // processNewTransaction processes a new transaction from the mempool
+// func (ts *TransactionStreamer) processNewTransaction(ctx context.Context, tx *types.Transaction) {
+// 	// Check if we already have this transaction
+// 	ts.mu.RLock()
+// 	_, exists := ts.txStore[tx.Hash().Hex()]
+// 	ts.mu.RUnlock()
+
+// 	if exists {
+// 		return
+// 	}
+
+// 	// Get full transaction details
+// 	_, _, err := ts.client.TransactionByHash(ctx, tx.Hash())
+// 	if err != nil {
+// 		fmt.Printf("Error getting transaction details: %v\n", err)
+// 		return
+// 	}
+
+// 	// Get sender address
+// 	sender, err := types.Sender(types.LatestSignerForChainID(tx.ChainId()), tx)
+// 	if err != nil {
+// 		fmt.Printf("Error getting sender address: %v\n", err)
+// 		return
+// 	}
+
+// 	// Create new stored transaction
+// 	storedTx := &StoredTransaction{
+// 		Hash: tx.Hash().Hex(),
+// 		Metadata: TransactionMetadata{
+// 			Status:       StatusQueued,
+// 			TimeReceived: time.Now().Unix(),
+// 			Type:         getTransactionType(tx),
+// 			Nonce:        tx.Nonce(),
+// 			From:         sender.Hex(),
+// 			To:           tx.To().Hex(),
+// 			Timestamp:    time.Now().Unix(),
+// 		},
+// 	}
+
+// 	// Set gas price based on transaction type
+// 	switch tx.Type() {
+// 	case types.DynamicFeeTxType:
+// 		storedTx.Metadata.MaxFeePerGas = tx.GasFeeCap()
+// 		storedTx.Metadata.MaxPriorityFee = tx.GasTipCap()
+// 	case types.BlobTxType:
+// 		storedTx.Metadata.MaxFeePerGas = tx.GasFeeCap()
+// 		storedTx.Metadata.MaxPriorityFee = tx.GasTipCap()
+// 		storedTx.Metadata.MaxFeePerBlobGas = tx.BlobGasFeeCap()
+// 	default:
+// 		storedTx.Metadata.GasPrice = tx.GasPrice()
+// 	}
+
+// 	// Store transaction
+// 	ts.mu.Lock()
+// 	ts.txStore[storedTx.Hash] = storedTx
+// 	ts.mu.Unlock()
+// }
+
+// // trackTransactionStatus tracks the status of transactions
+// func (ts *TransactionStreamer) trackTransactionStatus(ctx context.Context) {
+// 	ticker := time.NewTicker(5 * time.Second)
+// 	defer ticker.Stop()
+
+// 	for {
+// 		select {
+// 		case <-ticker.C:
+// 			ts.mu.RLock()
+// 			for _, tx := range ts.txStore {
+// 				// Skip if already mined or dropped
+// 				if tx.Metadata.Status == StatusMined || tx.Metadata.Status == StatusDropped {
+// 					continue
+// 				}
+
+// 				// Check if transaction is pending
+// 				if tx.Metadata.Status == StatusQueued {
+// 					var result map[string]map[string]map[string]*types.Transaction
+// 					err := ts.rpcClient.Call(&result, "txpool_content")
+// 					if err == nil {
+// 						// Check if transaction is in pending pool
+// 						for _, account := range result["pending"] {
+// 							if _, exists := account[tx.Hash]; exists {
+// 								ts.mu.RUnlock()
+// 								ts.mu.Lock()
+// 								tx.Metadata.Status = StatusPending
+// 								tx.Metadata.TimePending = time.Now().Unix()
+// 								ts.mu.Unlock()
+// 								ts.mu.RLock()
+// 								break
+// 							}
+// 						}
+// 					}
+// 				}
+
+// 				// Check if transaction is mined
+// 				if tx.Metadata.Status == StatusPending {
+// 					hash := common.HexToHash(tx.Hash)
+// 					receipt, err := ts.client.TransactionReceipt(ctx, hash)
+// 					if err == nil && receipt != nil {
+// 						ts.mu.RUnlock()
+// 						ts.mu.Lock()
+// 						tx.Metadata.Status = StatusMined
+// 						tx.Metadata.TimeMined = time.Now().Unix()
+// 						tx.Metadata.BlockNumber = receipt.BlockNumber.Uint64()
+// 						tx.Metadata.BlockHash = receipt.BlockHash.Hex()
+// 						ts.mu.Unlock()
+// 						ts.mu.RLock()
+// 					}
+// 				}
+// 			}
+// 			ts.mu.RUnlock()
+// 		case <-ctx.Done():
+// 			return
+// 		case <-ts.stopChan:
+// 			return
+// 		}
+// 	}
+// }
+
+// getTransactionType determines the type of transaction
+func getTransactionType(tx *types.Transaction) TransactionType {
+	switch {
+	case tx.Type() == types.BlobTxType:
+		return BlobTx
+	case tx.Type() == types.DynamicFeeTxType:
+		return EIP1559Tx
+	case tx.Type() == types.AccessListTxType:
+		return EIP2930Tx
+	default:
+		return LegacyTx
 	}
 }
